Homepage network model issue fix

+++ mobile/src/main/java/fit/asta/health/navigation/home/model/network/Weather.kt
@@ -39,7 +39,7 @@ data class CurrentWeather(
     @SerializedName("winddirection")
     val winddirection: Int,
     @SerializedName("windspeed")
-    val windspeed: Int
+    val windspeed: Double
 )
 
 data class Daily(
